# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omni - Documentação da API",
        "description": "Documentação de todos os endpoints da API do sistema Omni.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/site-zen": {
            "get": {
                "tags": [
                    "Site Zen"
                ],
                "summary": "Lista todos os registros de métricas do Site Zen",
                "operationId": "11a5d4a41a82374f03a827027c265a8c",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SiteZen"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Site Zen"
                ],
                "summary": "Cria um novo registro de métrica",
                "operationId": "fbdfcc67ed8bdd5229729adf00b40cdb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSiteZenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteZen"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/site-zen/{id}": {
            "get": {
                "tags": [
                    "Site Zen"
                ],
                "summary": "Busca um registro de métrica pelo ID",
                "operationId": "327515f4bb0cd4caccccdb408bfa3fae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteZen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Site Zen"
                ],
                "summary": "Atualiza um registro de métrica",
                "operationId": "14da63367f508edcab2607371be809db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSiteZenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SiteZen"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Site Zen"
                ],
                "summary": "Deleta um registro de métrica",
                "operationId": "7650420321db45a59184cd6ec16997fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso"
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tasks/{task}/activities": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Lista o histórico de atividades de uma tarefa",
                "operationId": "40938de65b8f337fee68f8825ce7a45b",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActivityLog"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreSiteZenRequest": {
                "required": [
                    "mes",
                    "visitas",
                    "compras",
                    "novas_visitas",
                    "taxa_conversao",
                    "ticket_medio",
                    "pa",
                    "receita_total"
                ],
                "properties": {
                    "mes": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "Janeiro"
                    },
                    "visitas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1500
                    },
                    "compras": {
                        "type": "integer",
                        "format": "int32",
                        "example": 50
                    },
                    "novas_visitas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1200
                    },
                    "taxa_conversao": {
                        "type": "number",
                        "format": "float",
                        "example": 3.33
                    },
                    "ticket_medio": {
                        "type": "number",
                        "format": "float",
                        "example": 250.75
                    },
                    "pa": {
                        "type": "number",
                        "format": "float",
                        "example": 1.5
                    },
                    "receita_total": {
                        "type": "number",
                        "format": "float",
                        "example": 12537.5
                    },
                    "produtos_mais_vistos": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "produtos_comprados": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "funil_usuarios": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateSiteZenRequest": {
                "properties": {
                    "mes": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "Janeiro"
                    },
                    "visitas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1500
                    },
                    "compras": {
                        "type": "integer",
                        "format": "int32",
                        "example": 50
                    },
                    "novas_visitas": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1200
                    },
                    "taxa_conversao": {
                        "type": "number",
                        "format": "float",
                        "example": 3.33
                    },
                    "ticket_medio": {
                        "type": "number",
                        "format": "float",
                        "example": 250.75
                    },
                    "pa": {
                        "type": "number",
                        "format": "float",
                        "example": 1.5
                    },
                    "receita_total": {
                        "type": "number",
                        "format": "float",
                        "example": 12537.5
                    },
                    "produtos_mais_vistos": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "produtos_comprados": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "funil_usuarios": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ActivityLog": {
                "title": "Activity Log Record",
                "description": "Registro de uma atividade (log) realizada em um modelo",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "log_name": {
                        "type": "string",
                        "example": "Task"
                    },
                    "description": {
                        "type": "string",
                        "example": "updated"
                    },
                    "subject_type": {
                        "type": "string",
                        "example": "App\\Models\\Task"
                    },
                    "subject_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "causer_id": {
                        "description": "ID do usuário que causou a ação",
                        "type": "integer",
                        "example": 1
                    },
                    "properties": {
                        "description": "JSON com os dados antigos e novos",
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "SiteZen": {
                "title": "Site Zen Analytics",
                "description": "Modelo que representa as métricas consolidadas de um site, geralmente agrupadas por mês.",
                "properties": {
                    "id": {
                        "description": "ID único do registro de métrica",
                        "type": "integer",
                        "readOnly": true
                    },
                    "mes": {
                        "description": "Mês de referência das métricas",
                        "type": "string",
                        "example": "Janeiro"
                    },
                    "visitas": {
                        "description": "Número total de visitas ao site no mês",
                        "type": "integer",
                        "example": 15000
                    },
                    "compras": {
                        "description": "Número total de compras realizadas no mês",
                        "type": "integer",
                        "example": 300
                    },
                    "novas_visitas": {
                        "description": "Número de visitantes que acessaram o site pela primeira vez",
                        "type": "integer",
                        "example": 9500
                    },
                    "taxa_conversao": {
                        "description": "Taxa de conversão de visitas para compras (%)",
                        "type": "number",
                        "format": "float",
                        "example": 2
                    },
                    "ticket_medio": {
                        "description": "Valor médio gasto por compra",
                        "type": "number",
                        "format": "float",
                        "example": 150.75
                    },
                    "pa": {
                        "description": "Métrica de performance 'PA' (Produtos por Atendimento/Atração)",
                        "type": "number",
                        "format": "float",
                        "example": 1.2
                    },
                    "receita_total": {
                        "description": "Receita total gerada no mês",
                        "type": "number",
                        "format": "float",
                        "example": 45225
                    },
                    "produtos_mais_vistos": {
                        "description": "Lista dos nomes dos produtos mais visualizados",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Produto A",
                            "Produto B",
                            "Produto C"
                        ]
                    },
                    "produtos_comprados": {
                        "description": "Lista dos nomes dos produtos mais comprados",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Produto A",
                            "Produto D"
                        ]
                    },
                    "funil_usuarios": {
                        "description": "Dados do funil de conversão de usuários",
                        "type": "array",
                        "items": {
                            "properties": {
                                "etapa": {
                                    "type": "string",
                                    "example": "Homepage"
                                },
                                "usuarios": {
                                    "type": "integer",
                                    "example": 15000
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}